requirements.txt
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.7.2
six==1.11.0
SQLAlchemy==1.2.6
Werkzeug==0.14.1
